# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "SalesERP API",
    "description": "Sales ERP Management System API",
    "version": "1.0.0"
  },
  "host": "erp-sales-backend.onrender.com",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "paths": {
...
    "/api/chart-of-accounts/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "item",
            "in": "query",
            "type": "string"
          },
          {
            "name": "nature",
            "in": "query",
            "type": "string"
          },
          {
            "name": "groupType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "format",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "parentGroup": {
                  "example": "any"
                },
                "item": {
                  "example": "any"
                },
                "nature": {
                  "example": "any"
                },
                "isSystemDefined": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/account-groups/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "chartOfAccountId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "branchId": {
                  "example": "any"
                },
                "chartOfAccountId": {
                  "example": "any"
                },
                "underGroup": {
                  "example": "any"
                },
                "groupName": {
                  "example": "any"
                },
                "shortName": {
                  "example": "any"
                },
                "gstin": {
                  "example": "any"
                },
                "pan": {
                  "example": "any"
                },
                "natureOfBusiness": {
                  "example": "any"
                },
                "creditPeriod": {
                  "example": "any"
                },
                "creditLimit": {
                  "example": "any"
                },
                "defaultPaymentMode": {
                  "example": "any"
                },
                "contact": {
                  "example": "any"
                },
                "openingBalance": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/bank-accounts/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          },
          {
            "name": "underGroup",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "branchId": {
                  "example": "any"
                },
                "underGroup": {
                  "example": "any"
                },
                "accountName": {
                  "example": "any"
                },
                "shortName": {
                  "example": "any"
                },
                "bankHolderName": {
                  "example": "any"
                },
                "accountNumber": {
                  "example": "any"
                },
                "ifsc": {
                  "example": "any"
                },
                "bankName": {
                  "example": "any"
                },
                "openingBalance": {
                  "example": "any"
                },
                "balanceType": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/items/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "propertyType",
            "in": "query",
            "type": "string"
          },
          {
            "name": "status",
            "in": "query",
            "type": "string"
          },
          {
            "name": "showInSales",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "branchId": {
                  "example": "any"
                },
                "itemCode": {
                  "example": "any"
                },
                "itemName": {
                  "example": "any"
                },
                "propertyType": {
                  "example": "any"
                },
                "projectName": {
                  "example": "any"
                },
                "location": {
                  "example": "any"
                },
                "area": {
                  "example": "any"
                },
                "dimensions": {
                  "example": "any"
                },
                "ratePerUnit": {
                  "example": "any"
                },
                "totalPrice": {
                  "example": "any"
                },
                "bookingAmount": {
                  "example": "any"
                },
                "status": {
                  "example": "any"
                },
                "floorNumber": {
                  "example": "any"
                },
                "totalFloors": {
                  "example": "any"
                },
                "bedrooms": {
                  "example": "any"
                },
                "bathrooms": {
                  "example": "any"
                },
                "parking": {
                  "example": "any"
                },
                "amenities": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "legalDetails": {
                  "example": "any"
                },
                "hsnCode": {
                  "example": "any"
                },
                "openingStock": {
                  "example": "any"
                },
                "currentStock": {
                  "example": "any"
                },
                "openingValue": {
                  "example": "any"
                },
                "showInSales": {
                  "example": "any"
                },
                "images": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/sales/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "paymentStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fromDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "customerName": {
                  "example": "any"
                },
                "customerContact": {
                  "example": "any"
                },
                "invoiceNo": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "subTotal": {
                  "example": "any"
                },
                "cgst": {
                  "example": "any"
                },
                "sgst": {
                  "example": "any"
                },
                "igst": {
                  "example": "any"
                },
                "tax": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "total": {
                  "example": "any"
                },
                "paymentStatus": {
                  "example": "any"
                },
                "amountPaid": {
                  "example": "any"
                },
                "balanceAmount": {
                  "example": "any"
                },
                "paymentMode": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/purchases/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "paymentStatus",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          },
          {
            "name": "fromDate",
            "in": "query",
            "type": "string"
          },
          {
            "name": "toDate",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "vendorName": {
                  "example": "any"
                },
                "vendorContact": {
                  "example": "any"
                },
                "purchaseNo": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "gstin": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "subTotal": {
                  "example": "any"
                },
                "cgst": {
                  "example": "any"
                },
                "sgst": {
                  "example": "any"
                },
                "igst": {
                  "example": "any"
                },
                "tax": {
                  "example": "any"
                },
                "discount": {
                  "example": "any"
                },
                "total": {
                  "example": "any"
                },
                "paymentStatus": {
                  "example": "any"
                },
                "amountPaid": {
                  "example": "any"
                },
                "balanceAmount": {
                  "example": "any"
                },
                "paymentMode": {
                  "example": "any"
                },
                "notes": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/expenses/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          },
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "to",
            "in": "query",
            "type": "string"
          },
          {
            "name": "paymentMethod",
            "in": "query",
            "type": "string"
          },
          {
            "name": "vendorId",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "branchId": {
                  "example": "any"
                },
                "vendorId": {
                  "example": "any"
                },
                "vendorName": {
                  "example": "any"
                },
                "voucherNo": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "gstin": {
                  "example": "any"
                },
                "mobileNo": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "billNo": {
                  "example": "any"
                },
                "billDate": {
                  "example": "any"
                },
                "paymentMethod": {
                  "example": "any"
                },
                "items": {
                  "example": "any"
                },
                "subtotal": {
                  "example": "any"
                },
                "taxAmount": {
                  "example": "any"
                },
                "total": {
                  "example": "any"
                },
                "specialNotes": {
                  "example": "any"
                },
                "expenseNo": {
                  "example": "any"
                },
                "description": {
                  "example": "any"
                },
                "amount": {
                  "example": "any"
                },
                "tax": {
                  "example": "any"
                },
                "accountHead": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/api/receipts/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "search",
            "in": "query",
            "type": "string"
          },
          {
            "name": "isActive",
            "in": "query",
            "type": "string"
          },
          {
            "name": "from",
            "in": "query",
            "type": "string"
          },
          {
            "name": "to",
            "in": "query",
            "type": "string"
          },
          {
            "name": "mode",
            "in": "query",
            "type": "string"
          },
          {
            "name": "accountId",
            "in": "query",
            "type": "string"
          },
          {
            "name": "depositTo",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "branchId": {
                  "example": "any"
                },
                "receiptNo": {
                  "example": "any"
                },
                "accountId": {
                  "example": "any"
                },
                "accountName": {
                  "example": "any"
                },
                "email": {
                  "example": "any"
                },
                "amountReceived": {
                  "example": "any"
                },
                "date": {
                  "example": "any"
                },
                "mode": {
                  "example": "any"
                },
                "referenceNo": {
                  "example": "any"
                },
                "depositTo": {
                  "example": "any"
                },
                "depositToName": {
                  "example": "any"
                },
                "enableGstOnAdvance": {
                  "example": "any"
                },
                "gstAmount": {
                  "example": "any"
                },
                "outstandingInvoices": {
                  "example": "any"
                },
                "totalPayment": {
                  "example": "any"
                },
                "totalDiscount": {
                  "example": "any"
                },
                "customerName": {
                  "example": "any"
                },
                "fromAccount": {
                  "example": "any"
                },
                "amount": {
                  "example": "any"
                },
                "specialNotes": {
                  "example": "any"
                },
                "isActive": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}